+++ src/hybrid/hybrid_search_result.c
@@ -61,7 +61,7 @@ void HybridSearchResult_Free(HybridSearchResult* result) {
  * Store a SearchResult from a source into the HybridSearchResult.
  * Updates the score of the SearchResult and marks the source as having results.
  */
-void HybridSearchResult_StoreResult(HybridSearchResult* hybridResult, SearchResult* searchResult, int sourceIndex) {
+void HybridSearchResult_StoreResult(HybridSearchResult* hybridResult, SearchResult* searchResult, size_t sourceIndex) {
   RS_ASSERT(hybridResult && searchResult);
   RS_ASSERT(sourceIndex < hybridResult->numSources);
